[TEST, NOP] Some clean up in chopper_pack and new tests for user errors.

+++ include/chopper/pack/pack_config.hpp
@@ -9,7 +9,7 @@ struct pack_config
     std::filesystem::path data_file;
     std::filesystem::path output_filename{"binning.out"};
     uint16_t t_max{64};
-    int aggregate_by_column{-1};
+    int8_t aggregate_by_column{-1};
     //!\brief If given, the hll sketches are dumped to this directory and restored when they already exist.
     std::filesystem::path hll_dir{};
     //!\brief The number of hash functions for the IBFs.
